One getDocAndPath function
