+++ libs/movie/src/lib/movie/form/media-videos/media-videos.component.html
@@ -39,8 +39,7 @@
       <file-upload
         [form]="activeVideo.ref"
         [storagePath]="getPath('otherVideos')"
-        allowedFileType="video"
-        filePrivacy="public">
+        allowedFileType="video">
           <h3 title>Upload Video</h3>
       </file-upload>
       <p class="mat-caption">Keep in mind that those videos will be <strong>public</strong> and <strong>viewable by anyone</strong> on the marketplace.</p>
