+++ libs/utils/src/lib/file-sanitizer.ts
@@ -2,6 +2,56 @@ export const privacies = ['public', 'protected'] as const;
 export type Privacy = typeof privacies[number];
 export const tempUploadDir = 'tmp';
 
+export function dissectFilePath(filePath: string | undefined) {
+  if (!filePath) {
+    throw new Error('Upload Error : Undefined File Path');
+  }
+
+  const filePathElements = filePath.split('/');
+
+  if (filePathElements.length < 4) {
+    const error = `Upload Error : File Path ${filePath} is malformed.`;
+    const solution = 'It should at least contain 3 slash.';
+    const example = 'Example: public/collection/id/field/fileName';
+    throw new Error(`${error} ${solution}\n${example}`);
+  }
+
+  // remove tmp/
+  let isInTmpDir = false;
+  if (filePathElements[0] === tempUploadDir) {
+    filePathElements.shift();
+    filePath = filePathElements.join('/');
+    isInTmpDir = true;
+  }
+
+  let security: string;
+  // remove "protected/"" or "public/"
+  if (privacies.includes(filePathElements[0] as any)) {
+    security = filePathElements.shift();
+  }
+
+  const collection = filePathElements.shift();
+  const docId = filePathElements.shift();
+
+  if (!docId || !collection) {
+    throw new Error('Invalid path pattern');
+  }
+
+  // remove the file name at the end
+  // `filePathElements` is now only composed by the field to update
+  filePathElements.pop();
+
+  const fieldToUpdate = filePathElements.join('.');
+
+  return {
+    isInTmpDir,
+    security,
+    collection,
+    docId,
+    fieldToUpdate
+  }
+}
+
 /**
  * Cleans filename ( before firestore upload for example )
  * @param str
